requirements.txt
Pillow==11.0.0
PyQt5==5.15.11
PyQt5_sip==12.16.1
